Improved ExpressionMatchers, etc.
ExpressionMatchers:
* added flatMap simplifier to some of the match results;
* fixed matchTotal;
Reorganized the specs by splitting off some into it directory.
