+++ src/test/scala/com/phasmidsoftware/number/misc/ContinuedFractionSpec.scala
@@ -7,7 +7,6 @@ import com.phasmidsoftware.number.misc.ConFrac.LongLazyListFrom
 import com.phasmidsoftware.number.misc.ContinuedFraction.{Hurwitz, fPiBy4Leibniz}
 import org.scalatest.flatspec
 import org.scalatest.matchers.should
-import org.scalatest.tagobjects.Slow
 
 class ContinuedFractionSpec extends flatspec.AnyFlatSpec with should.Matchers {
 
@@ -446,15 +445,6 @@ class ContinuedFractionSpec extends flatspec.AnyFlatSpec with should.Matchers {
     val qo = z.toRational(1E-10)
     qo should matchPattern { case Some(_) => }
     qo.get.toDouble shouldBe math.Pi +- 1E-9
-  }
-
-  /**
-    * NOTE: this representation of pi converges very slowly which is why we normally ignore it.
-    */
-  it should "define ContinuedFraction.FourOverPiLeibniz" taggedAs Slow in {
-    val z: ContinuedFraction = ContinuedFraction.FourOverPiLeibniz
-    val q: Rational = z.toRational(1000).invert
-    q.toDouble shouldBe math.Pi / 4 +- 2.5E-4
   }
 
   it should "define ContinuedFraction.PiSomayaji" in {
